VIH-9077 Update storage of eLinks people (#445)
* Add Leaver and LeftOn properties, only validate leavers have an Id property

* added missing parameters for test

* adding logic to not remove when person entry is null

+++ BookingsApi/BookingsApi.IntegrationTests/Helper/TestDataManager.cs
@@ -368,8 +368,11 @@ namespace BookingsApi.IntegrationTests.Helper
                 {
                     await using var db = new BookingsDbContext(_dbContextOptions);
                     var jp = await db.JudiciaryPersons.SingleOrDefaultAsync(x => x.ExternalRefId == externalRefId);
-                    db.JudiciaryPersons.Remove(jp);
-                    await db.SaveChangesAsync();
+                    if (jp != null)
+                    {
+                        db.JudiciaryPersons.Remove(jp);
+                        await db.SaveChangesAsync();    
+                    }
                     TestContext.WriteLine(@$"Remove Judiciary Person: {externalRefId}.");
                 }
                 catch (JudiciaryPersonNotFoundException)
